// lib/services/sanity/model/20250910/paragrafOptions.ts
export type Paragraf =
  | '11-2'
  | '11-3'
  | '11-4'
  | '11-5'
  | '11-6'
  | '11-7'
  | '11-8'
  | '11-9'
  | '11-10'
  | '11-11'
  | '11-12'
  | '11-13'
  | '11-14'
  | '11-15'
  | '11-16'
  | '11-17'
  | '11-18'
  | '11-19'
  | '11-20'
  | '11-21'
  | '11-22'
  | '11-23'
  | '11-24'
  | '11-25'
  | '11-26'
  | '11-27'
  | '11-28'
  | '11-29'
  | '11-30'
  | '11-31'

export type ParagrafOption = {
  title: string
  value: Paragraf
}

export const paragrafOptions: ParagrafOption[] = [
  {title: '11-2', value: '11-2'},
  {title: '11-3', value: '11-3'},
  {title: '11-4', value: '11-4'},
  {title: '11-5', value: '11-5'},
  {title: '11-6', value: '11-6'},
  {title: '11-7', value: '11-7'},
  {title: '11-8', value: '11-8'},
  {title: '11-9', value: '11-9'},
  {title: '11-10', value: '11-10'},
  {title: '11-11', value: '11-11'},
  {title: '11-12', value: '11-12'},
  {title: '11-13', value: '11-13'},
  {title: '11-14', value: '11-14'},
  {title: '11-15', value: '11-15'},
  {title: '11-16', value: '11-16'},
  {title: '11-17', value: '11-17'},
  {title: '11-18', value: '11-18'},
  {title: '11-19', value: '11-19'},
  {title: '11-20', value: '11-20'},
  {title: '11-21', value: '11-21'},
  {title: '11-22', value: '11-22'},
  {title: '11-23', value: '11-23'},
  {title: '11-24', value: '11-24'},
  {title: '11-25', value: '11-25'},
  {title: '11-26', value: '11-26'},
  {title: '11-27', value: '11-27'},
  {title: '11-28', value: '11-28'},
  {title: '11-29', value: '11-29'},
  {title: '11-30', value: '11-30'},
  {title: '11-31', value: '11-31'},
]